App Architecture Generator/requirements.txt
streamlit
google-generativeai
streamlit-mermaid
